Fix according to ShellCheck SC2115 warning
- https://www.shellcheck.net/wiki/SC2115

> Use "${var:?}" to ensure this never expands to /* .

+++ bin/setup_fonts.sh
@@ -33,7 +33,7 @@ do
   rm -rf "$TEMP_DIR"/"$fonts".zip
   if [ -d "$TEMP_DIR"/"$fonts" ]; then
     find "$TEMP_DIR"/"$fonts" -type f -name '*.ttf' -exec mv --target-directory="$USER_FONT_DIR" {} +
-    rm --recursive --force "$TEMP_DIR"/"$fonts"
+    rm --recursive --force "${TEMP_DIR:?}"/"$fonts"
   fi
   unset fonts
   unset font_family
